Fixes #8 #9 #10
--ESDRecordCustomerInvoice.cs
Correctly named setDefaultValuesForNullMembers function to be consistent with other order sale and order purchase classes.

--ESDRecordCustomerInvoiceLine.cs
Set correct decimal data type for properties storing pricing and decimals. Added properties to link products, downloads, and labour records from sales order, purchase order ans supplier invoice records.

--ESDRecordDeliveryNotice.cs
--ESDRecordDeliveryNoticeLine.cs
--ESDRecordInvoiceLineTax.cs
--ESDRecordInvoicePayment.cs
--ESDRecordInvoiceSurcharge.cs
--ESDRecordSupplierInvoice.cs
Set correct decimal data type for properties storing pricing and decimals.

--ESDRecordSupplierInvoiceLine.cs
Set correct decimal data type for properties storing pricing and decimals. Added properties to link products, downloads, and labour records from sales order, purchase order ans customer invoice records.

+++ Source/ESDRecordDeliveryNotice.cs
@@ -89,11 +89,11 @@ namespace EcommerceStandardsDocuments
 	
 	    /// <summary>Latitude co-ordinate of the geographic location where the goods are located at. The number must be between 90 and -90</summary>
         [DataMember(EmitDefaultValue = false)]
-        public double locationLatitude { get; set; }
+        public decimal locationLatitude { get; set; }
 
         /// <summary>longitude co-ordinate of the geographic location the goods are located at. The number must be between 180 and -180</summary>
         [DataMember(EmitDefaultValue = false)]
-        public double locationLongitude { get; set; }
+        public decimal locationLongitude { get; set; }
 	
 	    /// <summary>Text describing instructions on how to handle the order.</summary>
         [DataMember(EmitDefaultValue = false)]
